Refactor RuleEngine implementation to be more aligned with OpenClosed principle + Adding Database adapter project

+++ src/FraudDetection/Contracts/Ports/RuleEngineConfiguration.cs
@@ -5,6 +5,6 @@ namespace FraudDetection.Contracts.Ports
     public class RuleEngineConfiguration
     {
         public PersonComparisonRuleType RuleType{ get; set; }
-        public float Weight { get; set; }
+        public decimal Weight { get; set; }
     }
 }
